move marks start, end from Long -> Int

+++ feature_editor/src/main/java/com/agileburo/anytype/feature_editor/domain/Content.kt
@@ -9,8 +9,8 @@ sealed class Content {
 }
 
 data class Mark(
-    val start : Long,
-    val end : Long,
+    val start : Int,
+    val end : Int,
     val type : MarkType,
     val param: String
 ) {
